+++ tests/lib/rules/no-literal-string.js
@@ -33,7 +33,7 @@ function testFile(file) {
   };
 }
 
-ruleTester.run('no-literal-string', rule, {
+const usual = {
   valid: [
     testFile('valid.jsx'),
     { code: 'const a = "absfoo";<DIV abc="bcd" />' },
@@ -44,6 +44,14 @@ ruleTester.run('no-literal-string', rule, {
     {
       code: 'const a = "fooabc";',
       options: [{ mode: 'all', words: { exclude: ['^foo.*'] } }],
+    },
+    {
+      code: 'new Error("hello")',
+      options: [
+        {
+          ignoreCallee: ['Error'],
+        },
+      ],
     },
     {
       code: 'foo.bar("taa");',
@@ -129,14 +137,70 @@ ruleTester.run('no-literal-string', rule, {
       ],
       errors,
     },
-    // JSX
+    {
+      code: 'export const a = "hello_string";',
+      options: [{ message: 'Some error message' }],
+      errors: [{ message: 'Some error message' }],
+    },
+  ],
+};
+
+const jsx = {
+  valid: [
+    { code: '<div className="primary"></div>' },
+    { code: '<div className={a ? "active": "inactive"}></div>' },
+    { code: '<div>{i18next.t("foo")}</div>' },
+    { code: '<svg viewBox="0 0 20 40"></svg>' },
+    { code: '<line x1="0" y1="0" x2="10" y2="20" />' },
+    { code: '<path d="M10 10" />' },
+    {
+      code:
+        '<circle width="16px" height="16px" cx="10" cy="10" r="2" fill="red" />',
+    },
+    {
+      code:
+        '<a href="https://google.com" target="_blank" rel="noreferrer noopener"></a>',
+    },
+    {
+      code: '<div id="some-id" tabIndex="0" aria-labelledby="label-id"></div>',
+    },
+    { code: '<div role="button"></div>' },
+    { code: '<img src="./image.png" />' },
+    { code: '<A style="bar" />' },
+    { code: '<button type="button" for="form-id" />' },
+    { code: '<DIV foo="bar" />', options: [{ ignoreAttribute: ['foo'] }] },
+    { code: '<Trans>foo</Trans>' },
+    { code: '<Trans><span>bar</span></Trans>' },
+    { code: '<Trans>foo</Trans>', options: [{ ignoreComponent: ['Icon'] }] },
+    { code: '<Icon>arrow</Icon>', options: [{ ignoreComponent: ['Icon'] }] },
+    { code: '<div>{import("abc")}</div>', options: [{ markupOnly: true }] },
+    {
+      code: '<div>{[].map(item=>"abc")}</div>',
+      options: [{ markupOnly: true }],
+    },
+    { code: '<div>{"hello" + "world"}</div>', options: [{ markupOnly: true }] },
+    { code: '<DIV foo="FOO" />', options: [{ markupOnly: true }] },
+    {
+      code: '<DIV foo="bar" />',
+      options: [{ markupOnly: true, ignoreAttribute: ['foo'] }],
+    },
+    { code: '<DIV foo="bar" />', options: [{ onlyAttribute: ['bar'] }] },
+  ],
+  invalid: [
     {
       ...testFile('invalid-jsx-only.jsx'),
       options: [{ mode: 'jsx-only' }],
       errors: 5,
     },
+    {
+      code: '<div>{"hello world"}</div>',
+      options: [{ markupOnly: true, message: 'Some error message' }],
+      errors: [{ message: 'Some error message' }],
+    },
   ],
-});
+};
+ruleTester.run('no-literal-string', rule, usual);
+ruleTester.run('no-literal-string', rule, jsx);
 
 //
 // ─── VUE ────────────────────────────────────────────────────────────────────────
@@ -182,6 +246,22 @@ const tsTester = new RuleTester({
   },
 });
 
+tsTester.run('no-literal-string', rule, usual);
+tsTester.run(
+  'no-literal-string',
+  rule,
+  Object.entries(jsx).reduce(
+    (prev, [key, value]) => ({
+      ...prev,
+      [key]: value.map(item => ({
+        ...item,
+        filename: 'a.jsx',
+      })),
+    }),
+    {}
+  )
+);
+
 tsTester.run('no-literal-string', rule, {
   valid: [
     {
@@ -190,9 +270,7 @@ tsTester.run('no-literal-string', rule, {
     { code: '<div className="hello"></div>', filename: 'a.tsx' },
     { code: "var a: Element['nodeName']" },
     { code: "var a: Omit<T, 'af'>" },
-    { code: `var a: 'abc' = 'abc'` },
-    { code: `var a: 'abc' | 'name'  | undefined= 'abc'` },
-    { code: "type T = {name: 'b'} ; var a: T =  {name: 'b'}" },
+    { code: "function Button({ t= 'name'  }: {t: string}){} " },
     { code: "enum T  {howard=1, 'a b'=2} ; var a = T['howard']" },
     { code: "function Button({ t= 'name'  }: {t: 'name'}){} " },
     { code: "type T ={t?:'name'|'abc'};function Button({t='name'}:T){}" },
@@ -202,6 +280,12 @@ tsTester.run('no-literal-string', rule, {
       code: '<>foo123</>',
       filename: 'a.tsx',
       errors,
+    },
+    {
+      code: '<>foo999</>',
+      filename: 'a.tsx',
+      options: [{ markupOnly: true, message: 'Some error message' }],
+      errors: [{ message: 'Some error message' }],
     },
     {
       code: `<button className={styles.btn}>loading</button>`,