+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### 5.2.1 (2022-05-05)
+
+## 5.2.0 (2022-05-05)
+
+
+### Features
+
+* adding message as config option ([4ce1b18](https://github.com/edvardchen/eslint-plugin-i18next/commit/4ce1b1894af15c0f0e91f4094e527790322b8bc0))
+
+### 5.1.3 (2022-05-05)
+
+### 5.1.2 (2021-09-12)
+
+### [5.1.1](https://github.com/edvardchen/eslint-plugin-i18next/compare/v5.1.0...v5.1.1) (2021-04-08)
+
+
+### Bug Fixes
+
+* ignore strings as object property name ([e70bb8f](https://github.com/edvardchen/eslint-plugin-i18next/commit/e70bb8f)), closes [#42](https://github.com/edvardchen/eslint-plugin-i18next/issues/42)
+
 ## [5.1.0](https://github.com/edvardchen/eslint-plugin-i18next/compare/v5.0.0...v5.1.0) (2021-03-28)